+++ bambu_cli/camera.py
@@ -22,12 +22,7 @@ import urllib.error
 import urllib.request
 from urllib.parse import urlparse
 
-from bambu_cli.cli import (
-    _exception_for_message,
-    _expand_path,
-    _namespace_get,
-    _path_for_message,
-)
+from bambu_cli.argutils import namespace_get as _namespace_get
 from bambu_cli.config import load_access_code
 from bambu_cli.constants import (
     DEFAULT_NETWORK_TIMEOUT,
@@ -39,6 +34,9 @@ from bambu_cli.constants import (
 from bambu_cli.context import RuntimeContext
 from bambu_cli.errors import BambuError, abort
 from bambu_cli.logging_utils import logger, safe_log_error
+from bambu_cli.paths import exception_for_message as _exception_for_message
+from bambu_cli.paths import expand_path as _expand_path
+from bambu_cli.paths import path_for_message as _path_for_message
 from bambu_cli.utils import _ensure_parent_dir, emit_json, emit_json_error
 
 # The port-6000 camera stream's first frames can be stale (buffered from a